requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
crispy-bootstrap5==2025.6
Django==5.2.5
django-admin-thumbnails==0.2.9
django-cleanup==9.0.0
django-crispy-forms==2.4
django-parler==2.3
django-rosetta==0.9.9
django-widget-tweaks==1.5.0
idna==3.10
pillow==11.3.0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0